Dev 026 fixed response to client and time.Duration in config (#33)
"In this version, flexible support for configuration settings has been added to the configuration file, which allows for a more precise description of the necessary connection parameters for the client and the peers between which switching occurs. This MR was tested with 7 connected peers and one active client, which in real-time sent 1000 requests, during the tests there were no problems, the configuration settings used in the configuration file are provided in this build. If you need higher parameters, you can try to change the settings of the following keys: `clientDeadLineTime`, `destinationHostDeadLine`, `connectionSemaphore`, and `countDialAttemptsToPeer`, these keys are located in the config.yaml file. The output of results to the client was also corrected.